Mesh instance data transfer to shaders

+++ PathFinder/Source/Scene/Material.hpp
@@ -7,11 +7,11 @@ namespace PathFinder
 
     struct Material
     {
-        uint16_t AlbedoMapIndex = 0;
-        uint16_t NormalMapIndex = 0;
-        uint16_t RoughnessMapIndex = 0;
-        uint16_t MetalnessMapIndex = 0;
-        uint16_t AOMapIndex = 0;
+        uint32_t AlbedoMapIndex = 0;
+        uint32_t NormalMapIndex = 0;
+        uint32_t RoughnessMapIndex = 0;
+        uint32_t MetalnessMapIndex = 0;
+        uint32_t AOMapIndex = 0;
     };
 
 }
